Implement ClickHouse and DuckDB event stores with schema initialization and query capabilities
- Added ClickHouseEventsStore and DuckDbEventsStore implementations for event storage.
- Implemented methods for pinging, initializing schemas, inserting checkpoint events, and querying events.
- Introduced common database utility functions for escaping values and building SQL queries.
- Updated tests to ensure functionality of new event store implementations.
